Criação do cronograma assim que um projeto é cadastrado

+++ beelieve/src/main/java/com/api/beelieve/entidade/cronograma/dto/DadosMes.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record DadosMes(
 		String mes_cronograma,
-		Double ordem_mes_cronograma,
+		Integer ordem_mes_cronograma,
 		List<DadosPlanejamento> niveis
 		) {
